The onSearchBlur code was breaking the ability to click on a selection in the Instant search drop-down. Moved the code that shows and hides Instant search results into utils/search-functions, and we are now calling the "exitSearch" code once we arrive at Candidate or PositionList routes.
