+++ src/js/routes/Ballot/Candidate.jsx
@@ -9,6 +9,7 @@ import OfficeStore from "../../stores/OfficeStore";
 import PositionList from "../../components/Ballot/PositionList";
 import ThisIsMeAction from "../../components/Widgets/ThisIsMeAction";
 import VoterStore from "../../stores/VoterStore";
+import { exitSearch } from "../../utils/search-functions";
 
 export default class Candidate extends Component {
   static propTypes = {
@@ -29,6 +30,11 @@ export default class Candidate extends Component {
 
     this.listener = GuideStore.addListener(this._onChange.bind(this));
     GuideActions.retrieveGuidesToFollowByBallotItem(this.we_vote_id, "CANDIDATE");
+
+    // Display the candidate's name in the search box
+    var { candidate } = this.state;
+    var searchBoxText = candidate.ballot_item_display_name || "";  // TODO DALE Not working right now
+    exitSearch(searchBoxText);
   }
 
   componentWillUnmount () {
